# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlloPlan API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services/{serviceId}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message

  # ─── Packages ────────────────────────────────────────────────────────────────
  /services/{serviceId}/packages:
    get:
      operationId: listServicePackages
      tags: [packages]
      summary: List packages for a service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Packages list
          content:
            application/json:
              schema:
                type: object
                properties:
                  packages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Package"
                required:
                  - packages
    post:
      operationId: createServicePackage
      tags: [packages]
      summary: Create a package for a service
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageRequest"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{packageId}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package details
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageRequest"
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message

  # ─── Portfolio ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    VendorCard:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        businessType:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        verified:
          type: boolean
        subscriptionTier:
          type: string
          enum: [basic, pro, premium]
        coverImage:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        ownerName:
          type: string
        isFeatured:
          type: boolean
        isTopRated:
          type: boolean
        minPrice:
          type: number
          nullable: true
      required:
        - id
        - userId
        - reviewCount
        - verified
        - subscriptionTier
        - isAvailable
        - ownerName
        - isFeatured
        - isTopRated
...
    Service:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        basePrice:
          type: number
        isActive:
          type: boolean
        images:
          type: array
          items:
            type: string
        packagesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - name
        - category
        - basePrice
        - isActive
        - images
        - packagesCount
        - createdAt
...
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        basePrice:
          type: number
        images:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    Package:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        inclusions:
          type: array
          items:
            type: string
        durationHours:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - name
        - price
        - inclusions
        - isActive
        - createdAt
    PackageDetail:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        inclusions:
          type: array
          items:
            type: string
        durationHours:
          type: integer
          nullable: true
        isActive:
          type: boolean
        service:
          $ref: "#/components/schemas/Service"
        vendor:
          $ref: "#/components/schemas/VendorCard"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serviceId
        - name
        - price
        - inclusions
        - isActive
        - service
        - vendor
        - createdAt
    CreatePackageRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        inclusions:
          type: array
          items:
            type: string
        durationHours:
          type: integer
      required:
        - name
        - price
    UpdatePackageRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        inclusions:
          type: array
          items:
            type: string
        durationHours:
          type: integer
        isActive:
          type: boolean
...